fix: 解决代码pull下来之后，core/dist/下无编译的postisntall.js问题
